Crate syntastica_parsers_git

syntastica-parsers-git

Collection of tree-sitter parsers for syntastica, fetching with git in the build-script.

See the syntastica docs for more information on all parser collections.

List of included parsers

List of parsers included in the some feature

bash

c

cpp

css

go

html

java

javascript

json

python

rust

tsx

typescript

List of parsers additionally included in the most feature

asm

regex

List of parsers additionally included in the all feature

Features

At least one of some, most, and all must be enabled.

some — Include parsers for the most widely known supported languages.

most — Implies some. Include parsers for most common languages.

all — Implies most. Include parsers for all supported languages.

docs — Meant to be enabled when building docs

Structs

LanguageProviderImpl

An implementation of LanguageProvider, providing all parsers in the enabled feature set (see all) or a subset of them (see with_languages).

Functions

asmmost

Get the parser for asm.

bashsome

Get the parser for bash.

csome

Get the parser for c.

cppsome

Get the parser for cpp.

csssome

Get the parser for css.

gosome

Get the parser for go.

htmlsome

Get the parser for html.

javasome

Get the parser for java.

javascriptsome

Get the parser for javascript.

jsonsome

Get the parser for json.

pythonsome

Get the parser for python.

regexmost

Get the parser for regex.

rustsome

Get the parser for rust.

tsxsome

Get the parser for tsx.

typescriptsome

Get the parser for typescript.