syntastica_parsers_git

Enum Lang

#[non_exhaustive]
pub enum Lang {

    Asm,
    Bash,
    Bibtex,
    C,
    CSharp,
    Comment,
    Cpp,
    Css,
    Dart,
    Diff,
    Dockerfile,
    Ebnf,
    Ejs,
    Elixir,
    Erb,
    Fish,
    Go,
    Haskell,
    Hexdump,
    Html,
    Java,
    Javascript,
    Jsdoc,
    Json,
    Json5,
    Jsonc,
    Julia,
    Kotlin,
    Lalrpop,
    Latex,
    Llvm,
    Lua,
    Make,
    Markdown,
    MarkdownInline,
    Nix,
    Ocaml,
    OcamlInterface,
    Php,
    Printf,
    Python,
    Ql,
    Regex,
    Ruby,
    Rush,
    Rust,
    Scala,
    Scss,
    Sql,
    Toml,
    Tsx,
    Typescript,
    Typst,
    Ursa,
    Verilog,
    Wat,
    Yaml,
    Zig,

}

An enum of every supported language in the current feature set.

An instance of the respective tree-stter Language can be obtained with the get method.

You can also get a Lang from its name using for_name, or for a FileType using for_file_type. See the docs for each variant to see its “name” and the supported file types. Both of these require the SupportedLanguage trait to be in scope.

See LANGUAGES for a list containing all variants and LANGUAGE_NAMES for a list of all valid names.

The enum is marked as non-exhaustive for two reasons:

New languages may be added in the future

The variants are enabled/disabled by features

Example

use syntastica_parsers_git::{Lang, LANGUAGES, LANGUAGE_NAMES};
use syntastica_core::language_set::{SupportedLanguage, FileType};

// you can get a `Lang` from its name
assert_eq!(Lang::Rust, Lang::for_name("rust", &()).unwrap());
// and for a file type
assert_eq!(Some(Lang::Rust), Lang::for_file_type(FileType::Rust, &()));

// `LANGUAGES` is a list of all variants,
// `LANGUAGE_NAMES` is a list of all variant names
for (&lang, &name) in LANGUAGES.iter().zip(LANGUAGE_NAMES) {
    assert_eq!(lang, Lang::for_name(name, &()).unwrap());

    // `Lang` instances can be turned into strings
    assert_eq!(lang, Lang::for_name(<Lang as SupportedLanguage<'_, ()>>::name(&lang), &()).unwrap());
    assert_eq!(lang, Lang::for_name(lang.to_string(), &()).unwrap());
    assert_eq!(lang, Lang::for_name(lang.as_ref(), &()).unwrap());
    let lang_name: &'static str = lang.into();
    assert_eq!(lang, Lang::for_name(lang_name, &()).unwrap());
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Asm

Available on crate features most or asm only.

Provides the asm language, supports these file types: asm, vmasm.

Bash

Available on crate features some or bash only.

Provides the bash language, supports these file types: sh, bash, ksh.

Bibtex

Available on crate features all or bibtex only.

Provides the bibtex language, supports these file types: bib.

C

Available on crate features some or c only.

Provides the c language, supports these file types: c.

CSharp

Available on crate features most or c_sharp only.

Provides the c_sharp language, supports these file types: csharp.

Comment

Available on crate features most or comment only.

Provides the comment language, supports no file types.

Cpp

Available on crate features some or cpp only.

Provides the cpp language, supports these file types: cpp.

Css

Available on crate features some or css only.

Provides the css language, supports these file types: css.

Dart

Available on crate features most or dart only.

Provides the dart language, supports these file types: dart.

Diff

Available on crate features most or diff only.

Provides the diff language, supports these file types: diff.

Dockerfile

Available on crate features all or dockerfile only.

Provides the dockerfile language, supports these file types: dockerfile.

Ebnf

Available on crate features all or ebnf only.

Provides the ebnf language, supports these file types: ebnf.

Ejs

Available on crate features all or ejs only.

Provides the ejs language, supports these file types: ejavascript.

Elixir

Available on crate features most or elixir only.

Provides the elixir language, supports these file types: elixir.

Erb

Available on crate features all or erb only.

Provides the erb language, supports these file types: eruby.

Fish

Available on crate features all or fish only.

Provides the fish language, supports these file types: fish.

Go

Available on crate features some or go only.

Provides the go language, supports these file types: go.

Haskell

Available on crate features most or haskell only.

Provides the haskell language, supports these file types: haskell.

Hexdump

Available on crate features all or hexdump only.

Provides the hexdump language, supports these file types: hexdump.

Html

Available on crate features some or html only.

Provides the html language, supports these file types: html.

Java

Available on crate features some or java only.

Provides the java language, supports these file types: java.

Javascript

Available on crate features some or javascript only.

Provides the javascript language, supports these file types: javascript, jsx.

Jsdoc

Available on crate features most or jsdoc only.

Provides the jsdoc language, supports no file types.

Json

Available on crate features some or json only.

Provides the json language, supports these file types: json.

Json5

Available on crate features most or json5 only.

Provides the json5 language, supports these file types: json5.

Jsonc

Available on crate features most or jsonc only.

Provides the jsonc language, supports these file types: jsonc.

Julia

Available on crate features all or julia only.

Provides the julia language, supports these file types: julia.

Kotlin

Available on crate features some or kotlin only.

Provides the kotlin language, supports these file types: kotlin.

Lalrpop

Available on crate features all or lalrpop only.

Provides the lalrpop language, supports no file types.

Latex

Available on crate features all or latex only.

Provides the latex language, supports these file types: tex.

Llvm

Available on (crate features all or llvm) and non-target_family="wasm" only.

Provides the llvm language, supports these file types: llvm.

Lua

Available on crate features some or lua only.

Provides the lua language, supports these file types: lua.

Make

Available on crate features most or make only.

Provides the make language, supports these file types: make, automake.

Markdown

Available on crate features most or markdown only.

Provides the markdown language, supports these file types: markdown.

MarkdownInline

Available on crate features most or markdown_inline only.

Provides the markdown_inline language, supports no file types.

Nix

Available on crate features most or nix only.

Provides the nix language, supports these file types: nix.

Ocaml

Available on crate features all or ocaml only.

Provides the ocaml language, supports these file types: ocaml.

OcamlInterface

Available on crate features all or ocaml_interface only.

Provides the ocaml_interface language, supports these file types: ocamlinterface.

Php

Available on crate features most or php only.

Provides the php language, supports these file types: php.

Printf

Available on crate features most or printf only.

Provides the printf language, supports no file types.

Python

Available on crate features some or python only.

Provides the python language, supports these file types: python.

Ql

Available on crate features all or ql only.

Provides the ql language, supports these file types: ql.

Regex

Available on crate features most or regex only.

Provides the regex language, supports no file types.

Ruby

Available on crate features most or ruby only.

Provides the ruby language, supports these file types: ruby.

Rush

Available on crate features all or rush only.

Provides the rush language, supports these file types: rush.

Rust

Available on crate features some or rust only.

Provides the rust language, supports these file types: rust.

Scala

Available on crate features most or scala only.

Provides the scala language, supports these file types: scala, sbt.

Scss

Available on crate features most or scss only.

Provides the scss language, supports these file types: scss.

Sql

Available on crate features most or sql only.

Provides the sql language, supports these file types: sql, mysql.

Toml

Available on crate features some or toml only.

Provides the toml language, supports these file types: toml.

Tsx

Available on crate features some or tsx only.

Provides the tsx language, supports these file types: tsx.

Typescript

Available on crate features some or typescript only.

Provides the typescript language, supports these file types: typescript.

Typst

Available on crate features most or typst only.

Provides the typst language, supports these file types: typst.

Ursa

Available on crate features all or ursa only.

Provides the ursa language, supports these file types: ursa.

Verilog

Available on crate features all or verilog only.

Provides the verilog language, supports these file types: verilog.

Wat

Available on crate features all or wat only.

Provides the wat language, supports these file types: wat.

Yaml

Available on crate features some or yaml only.

Provides the yaml language, supports these file types: yaml.

Zig

Available on crate features most or zig only.

Provides the zig language, supports these file types: zig.

Implementations

impl Lang

pub fn get(&self) -> Language

Get an instance of the respective Language.

pub fn get_config(&self) -> Result<HighlightConfiguration>

Create an instance of the corresponding HighlightConfiguration.

pub fn highlights_query(&self) -> &'static str

Get the highlights query for this language.

pub fn injections_query(&self) -> &'static str

Get the injections query for this language.

pub fn locals_query(&self) -> &'static str

Get the locals query for this language.