Crate syntastica_parsers_git

syntastica-parsers-git

Collection of tree-sitter parsers for syntastica, fetching with git in the build-script.

See the project overview for more information on all parser collections.

List of included parsers

List of parsers included in the some feature

bash

c

cpp

css

go

html

java

javascript

json

python

rust

tsx

typescript

List of parsers additionally included in the most feature

asm

c_sharp

haskell

jsdoc

php

regex

ruby

scala

List of parsers additionally included in the all feature

ejs

erb

hexdump

julia

ocaml

ocaml_interface

ql

rush

verilog

wat

Features

At least one of some, most, and all must be enabled.

some — Include parsers for the most widely known supported languages.

most — Implies some. Include parsers for most common languages.

all — Implies most. Include parsers for all supported languages.

runtime-c (enabled by default) — Use the standard tree-sitter C runtime. See syntastica’s WebAssembly support for more information.

runtime-c2rust — Use the pure Rust tree-sitter runtime. See syntastica’s WebAssembly support for more information.

docs — Meant to be enabled when building docs

Structs

LanguageSetImpl

An implementation of [LanguageSet] including all languages in the enabled feature set.

Constants

LANGUAGES

A list of all language names that are supported by this parser collection.

Functions

asm

Get the parser for asm.

bash

Get the parser for bash. (not supported on the wasm32-unknown-unknown target)

c

Get the parser for c.

c_sharp

Get the parser for c_sharp.

cpp

Get the parser for cpp. (not supported on the wasm32-unknown-unknown target)

css

Get the parser for css.

ejs

Get the parser for ejs.

erb

Get the parser for erb.

go

Get the parser for go.

haskell

Get the parser for haskell.

hexdump

Get the parser for hexdump.

html

Get the parser for html. (not supported on the wasm32-unknown-unknown target)

java

Get the parser for java.

javascript

Get the parser for javascript.

jsdoc

Get the parser for jsdoc.

json

Get the parser for json.

julia

Get the parser for julia.

ocaml

Get the parser for ocaml. (not supported on the wasm32-unknown-unknown target)

ocaml_interface

Get the parser for ocaml_interface. (not supported on the wasm32-unknown-unknown target)

php

Get the parser for php.

python

Get the parser for python. (not supported on the wasm32-unknown-unknown target)

ql

Get the parser for ql.

regex

Get the parser for regex.

ruby

Get the parser for ruby. (not supported on the wasm32-unknown-unknown target)

rush

Get the parser for rush.

rust

Get the parser for rust.

scala

Get the parser for scala.

tsx

Get the parser for tsx.

typescript

Get the parser for typescript.

verilog

Get the parser for verilog.

wat

Get the parser for wat.